Many tweaks
