+++ apps/web-roo-code/src/components/homepage/index.ts
@@ -5,5 +5,4 @@ export * from "./faq-section"
 export * from "./features"
 export * from "./install-section"
 export * from "./testimonials"
-export * from "./user-logos"
 export * from "./whats-new-button"
